Client configuration libary - headers & docs
Initial sketch of a client configuration library. It's expected this
will subsume the client_setup functionality that we currently have, and
begin providing customizable load balancing support.

Releasing the headers now (no implementation work has been done) so that
we can bat around thoughts on this and make sure all the use cases are
implementable, and our primary use cases are easily so.


Former-commit-id: 9d0e047a5caba99194e6ed377ac0234f3907d6a3