Two fixes in grpc.framework.core._end
(1) Call "cancel" on each future, not on the list of futures.

(2) If and when futures mature their actions should simply abort all
outstanding operations and cancel any other futures. They should not
shut down the _End's internal thread pool; only the termination action
of the last operation to terminate should shut down the pool (in the
case of their having been active operations at the time at which the
_End's stop(grace) method was called).


Former-commit-id: 4b41ba249df4f290b3551b82b6004a0ce77bf2cf