Make servers and stubs context managers
Servers and stubs were context managers in the Alpha API; they may not
need to be in the Beta API but it's easy enough to do, eases migration,
and probably helps some use cases.

For now the stub is given empty __enter__ and __exit__ methods; we can
always come back and implement the actual work of context management in
a later change.


Former-commit-id: c824eb4e90fe150b113b2ceaa14fa4105718cc46