Reintroduce DNS-based cases to dualstack_socket_test.c.
These were dropped before the initial release because they depended on
a hostname that was only resolvable inside Google.  The newer
unittest.grpc.io domain is publicly accessible.

Allow the tests to be skipped, just in case someone is running an
isolated machine, or a DNS resolver that doesn't permit external domains
to return loopback addresses.


Former-commit-id: 6a70b012391a1c17f85cde6e02f80337e2913d69