Split AllTests into 4 schemes
For local running one can still tests AllTests, but for C.I. one can
test the 4 of them in parallel (and, importantly, deactivate the remote
interop tests until we fix them on Travis; while keeping the rest).

I had to make the remote tests a subclass of an abstract InteropTests
(like InteropTestsLocalSSL and InteropTestsLocalCleartext), and then
deactivate the abstract InteropTests on the 4 schemes that use it.

Former-commit-id: 91db5acd5b52a41d7069fd56bf0f6802b6c6e4b7