Migrate _termination_action into _End
For a termination action to properly mark an _End as having stopped it
must clear the _End's _cycle attribute. To be able to do that the
termination action must hold a reference to the _End instance.
Migrating the _termination_action behavior that creates termination
actions into the scope of the _End instance is the best way to afford
the needed instance reference.


Former-commit-id: e0abeeaa6d360a35d6a66da16d047fd2abc2a5d7