Delay reading stdout from disk until process terminates
The current code reads the stdout of each running process each time it
goes through a reap() iteration, and throws it away if the process is
still running.

On Windows, where we poll for completion every 100ms, I expect this is
having a serious impact on Jenkins performance, especially for noisier
tests.


Former-commit-id: db21899edcd663bf941b0c7ff07d5a3218908c9b