Add HTTP request parsing.
This extends the existing http parser to support requests as well as responses.

httpcli continues to exist and work as it has previously, though in the new
directory src/core/http (to reflect the fact the directory now contains code
relevant to parsing requests, which httpcli would not generally involve itself
in).


Former-commit-id: 1824f0519f26b00cb5e95e52ea50c7990223157c