Don't allow precompiled binaries in tests
Before it was possible for precompiled binaries of the Cython layer to
be downloaded during a test run if the metadata of the host machine's
gRPC Python project matched the metadata of a file in cloud storage. Now
we disallow this, and fix a bug where the relevant environment variable
was ineffectual.


Former-commit-id: 7d8ac466d6562dd6ae985de9ec275c50fa7fcfed