Correctly handle lfWidth in GDI.
The LOGFONT::lfWidth field is related to the LOGFONT::lfHeight field,
it is not related in any way to the OS/2::usWidthClass (unlike
LOGFONT::lfWeight, which mirrors the OS/2::usWeightClass). Be sure to
canonicalize this field to 0 (default) and not use it as a style value.

GOLD_TRYBOT_URL= https://gold.skia.org/search2?unt=true&query=source_type%3Dgm&master=false&issue=1819973003

Review URL: https://codereview.chromium.org/1819973003


Former-commit-id: 59f093d5d832a83b409879de287745e99f75ebfe