Make BoringSSL work with frameworks
Cocoapods 1.0 keeps the directory structure of the public headers when
creating a dynamic framework, so if we keep the header_mappings_dir as it
was, includes would need to be of the form #include <openssl/openssl/ssl.h>

This means our header_mappings_dir has to be ‘include/openssl’ instead of
‘include’. Which in turn means that, for static libraries, we have to tell
Cocoapods to prepend an ‘openssl’ directory to the headers. We do that
with the ‘header_dir’ attribute of the podspec.


Former-commit-id: 5842277ecbbbe694a828bd51934a28673f2f0a7f