Update SK_IMAGE_VERSION, run nanobench on its own image set
I've added a new directory to the cloud
(gs://chromium-skia-gm/skimage/v4/nanobench) which will
allow us to run dm and nanobench on different sets of
images.

The dm image set will stay the same, and includes
incomplete images, edge cases images, etc.

The nanobench set is much smaller, and includes jpegs and
pngs that we've found useful to bench in the past, plus
a couple of each for the other formats.  We can refine
this set as we move forward (and may also want to monitor
these images on perf.skia.org).

For now the goal is to reduce nanobench runtime by not
testing on many, many obscure images that are not
particularly interesting to time.

The actual change to use nanobench images is in:
https://codereview.chromium.org/1917903003/

BUG=skia:5181
GOLD_TRYBOT_URL= https://gold.skia.org/search2?unt=true&query=source_type%3Dgm&master=false&issue=1913813004

Review URL: https://codereview.chromium.org/1913813004


Former-commit-id: 0b55664afc854b03082c467505b1a811721bcd7a