Delete unused code from SkColorSpace
IMO, a lot of this code reflects an oversimplified view of
handling color profiles.  And a lot of this code is valid
math/checks/enums that we will eventually need.

Right now, it's all unused and is adding a bit of clutter
and confusion as I try to redesign this object.  I think
it makes sense to start simpler and build this back up as
necessary.

BUG=skia:
GOLD_TRYBOT_URL= https://gold.skia.org/search2?unt=true&query=source_type%3Dgm&master=false&issue=1932743002

Review-Url: https://codereview.chromium.org/1932743002

Former-commit-id: 23bd7e9273438863149ed58c089e9bf457e9892d