Moved grpc_shutdown to end of Py_Finalize()
We currently rely on the __del__ method of a module scope object
to call grpc_shutdown().  __del__ methods are not guaranteed
to be called, and furthermore there are no guarantees about
ordering, leading to shutdown race conditions.  This moves
grpc_shutdown to Py_Finalize(), which gets called after the
Python context is completely cleaned up.


Former-commit-id: f4c1bff0d8caedf716b460c4ee6114ac126e8872