Updating FeatureList default initialization pattern
FeatureList currently expects a singleton to be manually set by all code
which uses it.  This works fine in the browser where it is intialized
and in unit tests which use a default initializer, however for code
outside the browser which relies on the various components in Chromium,
this causes their code to crash.

An example of this is CL:
https://codereview.chromium.org/2101283004/

This was an innocuous change in a component which CRD relied on.
Since the unit tests are set up to never catch these kind of problems,
our component started crashing and we had to investigate to figure out
the root cause.

This change updates the feature and trial checks to initialize the
singleton if it was not already initialized.  If a caller tries to
re-initialize the singleton after that, the init code will DCHECK as
this represents a race condition and we want to ensure consistency for
the callers who rely on these checks being accurate.

BUG=624879

Review-Url: https://codereview.chromium.org/2110083007
Cr-Commit-Position: refs/heads/master@{#404243}
(cherry picked from commit 9dd1ba8a69f93aabeecfa00be85fe9cb76b871a9 [formerly 958f0473b8039a057b0e31412c2acae7dfda39cd])

Review URL: https://codereview.chromium.org/2138923002 .

Cr-Commit-Position: refs/branch-heads/2785@{#85}
Cr-Branched-From: 1ac8cbeb262a801827381e10bb796676610ddf00 [formerly 68623971be0cfc492a2cb0427d7f478e7b214c24]-refs/heads/master@{#403382}

Former-commit-id: 8ddcefe1e6ea7e133782e06b27e9767c94755840