Use dFdx in Vulkan to address distance field issues.
On Nexus Player and occasionally Nexus 5x we get transparent boxes around
paths. This appears to be because the dFdy call is not as accurate as
dFdx, which is the opposite of Mali 400. As Mali 400 is not supported with
Vulkan, we can go back to using dFdx in this case.

BUG=skia:5523
GOLD_TRYBOT_URL= https://gold.skia.org/search?issue=2163213004

Review-Url: https://codereview.chromium.org/2163213004

Former-commit-id: e499adf328bd7fc15a755325749f54c2b7e71f54