Add gpr_slice_new_with_user_data.
gpr_slice_new_with_user_data is like gpr_slice_new, but allows the
caller to specify a distinct pointer to pass to the destroy function.
This is useful when the data is part of a larger data structure that
should be destroyed when the data is no longer needed.


Former-commit-id: fca59fb2aa4a0be093d12162c763b8563ada4087