UPSTREAM:list: fix an issue with android build using clang
Sorry, I don't understand the android build system enough to say *which*
version of clang this effects, but either "clang-2812033" or
"clang-3016494" (probably the later).

But when 'sample' is undefined (ie. unitialized variable), the result is
not as well defined as it is with gcc.  Instead use a typeof() cast with
a defined value (ie. zero).

This fixes a crash that was reported on android.

Reported-by: John Stultz <john.stultz@linaro.org>
Signed-off-by: Rob Clark <robclark@freedesktop.org>
(cherry picked from commit b214b05ccd433c484a6a65e491a1a51b19e4811d)
Signed-off-by: Akshu Agrawal <akshu.agrawal@amd.com>

Change-Id: I7743bb16cc2b03dc8fc22f8b2ef9bcea96cd78d9
Reviewed-on: https://chromium-review.googlesource.com/412590
Commit-Ready: Akshu Agrawal <akshu.agrawal@amd.com>
Tested-by: Akshu Agrawal <akshu.agrawal@amd.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>