Avoid static initialization of the kIdentity block
Unlike other Objective-C objects, there's no hard reason why the compiler wouldn't be able to initialize a block statically (as it does with NSString literals). And it certainly doesn't complain about it (like it does with other object initializers). But as I haven't been able to find confirmation of this, and we're seeing a weird crash occur near this code, let's play it safe.

Former-commit-id: ea5325c484571cb114946b6e1d06e4a26284e198