[M53] Disable CMAA for samus
Selective enabling for CMAA on samus was added in https://crrev.com/402913 by
using a workaround with an exception to turn it on.

It was found that this caused a graphics corruption issue with the Camera app.
This CL is the minimal change to disable the feature by removing the samus PCI
ID from the list of exceptions to the workaround which disables CMAA.

BUG=642290
TEST=Bisecter reported reverting this commit resolved corruption issues.
NOPRESUBMIT=true
NOTRY=true

Review-Url: https://codereview.chromium.org/2296173002
Cr-Commit-Position: refs/branch-heads/2785@{#812}
Cr-Branched-From: 1ac8cbeb262a801827381e10bb796676610ddf00 [formerly 68623971be0cfc492a2cb0427d7f478e7b214c24]-refs/heads/master@{#403382}

Former-commit-id: 3ec0a76c3c48338c08bf489c442f69c2c03d18fd