Fix for Chrome MSAN
I believe that the complaints are occurring because the |a| vector
might be uninitialized where it is used here.  It doesn't actually
matter because we won't use or store that value - it's just a
placeholder.

But we need to make the bot happy.

BUG=skia:

GOLD_TRYBOT_URL= https://gold.skia.org/search?issue=3800

Change-Id: I1891da9d1d2708008e4606daebf9bb6f96e92fc0
Reviewed-on: https://skia-review.googlesource.com/3800
Commit-Queue: Matt Sarett <msarett@google.com>
Reviewed-by: Mike Klein <mtklein@chromium.org>

Former-commit-id: a9f64dec63b6f4169ba16f8b6c63fff5a6494029