Always use a color table with 256 colors
Speculative fix for skbug.com/5883

We are hitting an assert (which I have not been able to repro - still
need to get the skp that triggers it) because we are trying to read
beyond the end of the color table. We use 8-bit indices, so 256 should
be enough to prevent going beyond the end. There is only one case when
we use a smaller color table - when it is a dummy. Make the dummy full
size.

NOTREECHECKS=true

BUG=skia:5883
GOLD_TRYBOT_URL= https://gold.skia.org/search?issue=2452673002

Review-Url: https://codereview.chromium.org/2452673002

Former-commit-id: 0057ac11fccc82bc5b0c1bb0aefe838cc04efd13